+++ TractorMod/release-notes.md
@@ -1,6 +1,7 @@
 # Release notes
 ## Upcoming release
 * Fixed chests being treated as weeds in some cases.
+* Fixed scythe and some melee weapons not clearing dead crops if enabled.
 * Improved translations. Thanks to Mysti57 (added French)!
 
 ## 4.8